requirements.txt
amqp==2.3.2
anyjson==0.3.3
Babel==2.6.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1
django-heroku==0.3.1
django-modeltranslation==0.12.2
django-payeer==0.4
flower==0.9.2
idna==2.7
kombu==4.2.1
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
six==1.11.0
tornado==5.1
urllib3==1.23
vine==1.1.4
whitenoise==4.0
gunicorn